refactor: create carrier folder and move modules
