+++ powersimdata/network/constants/carrier/color.py
