+++ powersimdata/network/constants/carrier/efficiency.py
