+++ powersimdata/network/constants/carrier/emission.py
