+++ powersimdata/network/constants/carrier/resource.py
