+++ powersimdata/network/constants/region/zones.py
@@ -1,6 +1,7 @@
 from importlib import import_module
 
 from powersimdata.network.constants.model import model2region
+from powersimdata.network.helpers import check_model
 
 
 def get_zones(interconnect, model):
@@ -10,6 +11,8 @@ def get_zones(interconnect, model):
     :param str model: the grid model.
     :return: (*func*) -- function returning information on zones for a given model.
     """
+    check_model(model)
+
     mod = import_module(
         f"powersimdata.network.constants.region.{model2region[model].lower()}"
     )
